@@ -19,45 +19,43 @@
     "mongo:start": "net start MongoDB",
     "mongo:stop": "net stop MongoDB",
     "mongo:remove": "\"C:\\MongoDB\\mongod.exe\" — remove",
+    "mongo:path": "sudo service mongod stop && sudo mongod --dbpath /media/tuanmjnh/data/database/MongoDB/data",
     "eslint:init": "eslint --init"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "body-parser": "^1.19.0",
+    "axios": "^0.21.4",
+    "body-parser": "^1.19.1",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
-    "dotenv": "^8.2.0",
+    "dotenv": "10.0.0",
     "express": "^4.17.1",
     "express-flash": "0.0.2",
-    "express-session": "^1.17.0",
-    "formidable": "^1.2.2",
+    "express-session": "^1.17.2",
+    "formidable": "^1.2.6",
     "https-proxy-agent": "^5.0.0",
     "jsonwebtoken": "^8.5.1",
-    "lusca": "^1.6.1",
-    "moment": "^2.24.0",
-    "mongoose": "^5.12.13"
+    "lusca": "^1.7.0",
+    "moment": "^2.29.1",
+    "mongoose": "^6.1.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/preset-env": "^7.14.4",
+    "@babel/core": "^7.16.0",
+    "@babel/preset-env": "^7.16.4",
     "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.2.2",
+    "babel-loader": "^8.2.3",
     "babel-polyfill": "^6.26.0",
-    "concurrently": "^5.3.0",
+    "concurrently": "^6.4.0",
     "copy-webpack-plugin": "^6.4.1",
     "errorhandler": "^1.5.1",
-    "eslint": "^6.8.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-loader": "^3.0.2",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-node": "^10.0.0",
-    "eslint-plugin-promise": "^4.3.1",
-    "eslint-plugin-standard": "^4.1.0",
-    "html-webpack-plugin": "^5.3.1",
-    "nodemon": "^2.0.7",
-    "webpack": "^5.38.1",
-    "webpack-cli": "^3.3.12",
+    "eslint": "^8.6.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "html-webpack-plugin": "^5.5.0",
+    "nodemon": "^2.0.15",
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1",
     "webpack-node-externals": "^1.7.2"
   },
   "node": {
