@@ -27,30 +27,32 @@
     "upgrade-quasar:install-linux": "~/.npm-global/bin/quasar upgrade -i"
   },
   "dependencies": {
-    "@quasar/extras": "^1.10.6",
-    "axios": "^0.21.1",
-    "chart.js": "^3.3.2",
-    "core-js": "^3.14.0",
+    "@j-t-mcc/vue3-chartjs": "^1.2.0",
+    "@quasar/cli": "^1.2.2",
+    "@quasar/extras": "^1.12.2",
+    "axios": "^0.21.4",
+    "chart.js": "^3.6.2",
+    "core-js": "^3.19.3",
     "js-cookie": "^2.2.1",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
     "nprogress": "^0.2.0",
-    "qrcode": "^1.4.4",
-    "quasar": "^2.0.0-rc.1",
-    "vue-i18n": "^9.1.6",
+    "qrcode": "^1.5.0",
+    "quasar": "^2.4.3",
+    "vue-i18n": "^9.1.9",
     "vuedraggable": "^2.24.3",
-    "vuex": "^4.0.1",
-    "xlsx": "^0.16.9"
+    "vuex": "^4.0.2",
+    "xlsx": "^0.17.4"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.14.4",
-    "@intlify/vue-i18n-loader": "^2.1.0",
-    "@quasar/app": "^3.0.0-rc.1",
+    "@babel/eslint-parser": "^7.16.3",
+    "@intlify/vue-i18n-loader": "^2.1.2",
+    "@quasar/app": "^3.2.5",
     "copy-webpack-plugin": "^8.1.1",
-    "eslint": "^7.28.0",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-vue": "^7.10.0",
-    "eslint-webpack-plugin": "^2.4.0",
+    "eslint-plugin-vue": "^7.20.0",
+    "eslint-webpack-plugin": "^2.6.0",
     "quasar-app-extension-qdraggabletree": "0.0.5"
   },
   "browserslist": [
