@@ -1,7 +1,9 @@
 const mongoose = require('mongoose')
 const schema = new mongoose.Schema({
   // type: { type: String, default: null },
-  categories: { type: mongoose.Schema.Types.ObjectId, ref: 'categories', required: true },
+  categories: { type: Array, default: null },
+  // categories: { type: mongoose.Schema.Types.ObjectId, ref: 'categories', required: true },
+  // subCategory: { type: mongoose.Schema.Types.ObjectId, ref: 'subCategory', required: true },
   title: { type: String, required: true },
   code: { type: String, required: true, uppercase: true },
   desc: { type: String, default: null },
@@ -16,10 +18,12 @@ const schema = new mongoose.Schema({
   unit: { type: String, default: null }, // { type: mongoose.Schema.Types.ObjectId, ref: 'types' },
   origin: { type: String, default: null },
   date: { type: String, default: null },
-  pin: { type: Array, default: null },
+  pins: { type: Array, default: null },
   tags: { type: Array, default: null },
   attr: { type: Array, default: null },
   meta: { type: Array, default: null },
+  imei: { type: Number, default: null },
+  qrcode: { type: Number, default: null },
   // start_at: { type: Date, default: null },
   // end_at: { type: Date, default: null },
   order: { type: Number, default: 1 },
