@@ -2,10 +2,10 @@ const mongoose = require('mongoose')
 const schema = new mongoose.Schema({
   // type: { type: Number, required: true }, // 1 Import 2 Export
   code: { type: String, required: true, lowercase: true },
-  product: { type: Number, default: 0 },
-  quantity: { type: Number, default: 0 },
-  price: { type: Number, default: 0 },
-  vat: { type: Number, default: 0 },
+  products: { type: Number, default: 0 },
+  quantities: { type: Number, default: 0 },
+  prices: { type: Number, default: 0 },
+  taxes: { type: Number, default: 0 },
   createdAt: { type: Date, default: new Date() },
   createdBy: { type: mongoose.Schema.Types.ObjectId, ref: 'users' },
   createdIp: { type: String, default: null },
