@@ -16,19 +16,16 @@ String.prototype.convertToAscii = function () {
 String.prototype.removeChars = function () {
   return this.replace(/[~`!@#$%^&*()\[{}\]\\|;:\'\",<>./?]/g, '')
 }
-
 String.prototype.toHtml = function () {
   if (!this) return this
   var el = document.createElement('div')
   el.innerHTML = this
   return el.firstChild.data
 }
-
 String.prototype.toUpperCaseFirst = function () {
   if (!this) return this
   return this.charAt(0).toUpperCase() + this.slice(1)
 }
-
 String.prototype.toUpperCaseSpace = function () {
   if (!this) return this
   const arr = this.trim().split(' ')
@@ -38,6 +35,35 @@ String.prototype.toUpperCaseSpace = function () {
     if (i < arr.length - 1) rs += ' '
   }
   return rs
+}
+// String.prototype.formatDate = function (format = 'DD/MM/YYYY') {
+//   if (!this) return this
+//   return moment(this).format(format)
+// }
+
+// Number
+Number.prototype.humanFileSize = function (si = false, dp = 1) {
+  let bytes = this
+  const thresh = si ? 1000 : 1024
+
+  if (Math.abs(bytes) < thresh) return bytes + ' B'
+
+  const units = si ? ['KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'] : ['KiB', 'MiB', 'GiB', 'TiB', 'PiB', 'EiB', 'ZiB', 'YiB']
+  let u = -1
+  const r = 10 ** dp
+
+  do {
+    bytes /= thresh
+    ++u
+  } while (Math.round(Math.abs(bytes) * r) / r >= thresh && u < units.length - 1)
+
+  return bytes.toFixed(dp) + ' ' + units[u]
+}
+Number.prototype.toThousandFilter = function (num) {
+  return (+num || 0).toString().replace(/^-?\d+/g, m => m.replace(/(?=(?!\b)(\d{3})+$)/g, ','))
+}
+Number.prototype.NumberFormat = function (language = 'en-US') {
+  return new Intl.NumberFormat(language).format(this)
 }
 
 // Array
@@ -58,7 +84,6 @@ Array.prototype.pushIfNotExist = function (element, key) {
     }
   }
 }
-
 Array.prototype.pushIfNotExistUpdate = function (element, key) {
   if (Array.isArray(element)) {
     element.forEach(e => {
@@ -87,7 +112,6 @@ Array.prototype.pushIfNotExistUpdate = function (element, key) {
     }
   }
 }
-
 Array.prototype.sum = function (prop) {
   var total = 0
   for (var i = 0, _len = this.length; i < _len; i++) {
