@@ -1,11 +1,12 @@
 <template>
-  <div id="tm-print" :style="hidden?'opacity:0;width:0;height:0;':''">
+  <div ref="tmPrinter" :style="hidden?'display: none;':''">
+    <!--opacity:0;width:0;height:0;-->
     <slot name="content"></slot>
   </div>
 </template>
 
 <script>
-import { defineComponent } from 'vue';
+import { defineComponent, ref } from 'vue';
 export default defineComponent({
   name: 'tm-printer',
   props: {
@@ -15,29 +16,32 @@ export default defineComponent({
     // height: { type: Number, default: 900 },
     // timeout: { type: Number, default: 1000 }
   },
-  setup (props) {
-    const onExc = ({ styles, width, height, timeout }) => {
+  setup () {
+    const tmPrinter = ref(null)
+    // const idElement = 'id' + (new Date()).getTime()
+    const onExc = ({ title, styles, width, height, timeout }) => {
       width = width || 800
-      height = height || 900
+      height = height || 600
       timeout = timeout || 1000
       // Get HTML to print from element
-      const prtHtml = document.getElementById('tm-print').innerHTML
+      // const prtHtml = document.getElementById('tm-print').innerHTML
       // Get all stylesheets HTML
       let stylesHtml = ''
       for (const node of [...document.querySelectorAll('link[rel="stylesheet"], style')]) {
         stylesHtml += node.outerHTML
       }
-      stylesHtml += styles
+      stylesHtml += styles ? styles : ''
       // Open the print window
       const WinPrint = window.open('', '', `left=0,top=0,width=${width},height=${height},toolbar=0,scrollbars=0,status=0`)
 
       WinPrint.document.write(`<!DOCTYPE html>
         <html>
           <head>
-            ${stylesHtml}
+          <title>${title ? title : 'Printer'}</title>
+            ${stylesHtml ? stylesHtml : ''}
           </head>
           <body>
-            ${prtHtml}
+            ${tmPrinter.value.innerHTML}
           </body>
         </html>`)
       setTimeout(() => {
@@ -45,9 +49,10 @@ export default defineComponent({
         WinPrint.focus()
         WinPrint.print()
         WinPrint.close()
+        // prtHtml.innerHTML = ''
       }, timeout)
     }
-    return { props, onExc }
+    return { tmPrinter, onExc }
   }
 })
 </script>
