@@ -3,9 +3,15 @@
 </template>
 
 <script>
-export default {
-
-}
+import { defineComponent, defineAsyncComponent, ref, computed } from 'vue';
+export default defineComponent({
+  name: 'OrderIndex',
+  components: {
+  },
+  setup () {
+    return {}
+  }
+})
 </script>
 
 <style>
