@@ -17,7 +17,7 @@ const constant = {
   unit: null,
   origin: null,
   date: null,
-  pin: [],
+  pins: [],
   tags: null,
   attr: [],
   meta: null,
@@ -102,7 +102,7 @@ const actions = {
     return api.post(collection, params).then((res) => {
       commit('ADD_ITEMS', res)
       commit('SET_ITEM')
-      return (res)
+      return res
     })
   },
   put ({ commit }, params) {
