@@ -1,8 +1,8 @@
-function sprintf() {
+function sprintf () {
   var args = arguments,
     string = args[0],
     i = 1
-  return string.replace(/%((%)|s|d)/g, function(m) {
+  return string.replace(/%((%)|s|d)/g, function (m) {
     // m is the matched format, e.g. %s, %d
     var val = null
     if (m[2]) {
@@ -24,7 +24,7 @@ function sprintf() {
   })
 }
 
-String.prototype.format = function() {
+String.prototype.format = function () {
   if (arguments.length > 0 && typeof arguments[0] === 'object') {
     let $this = this
     arguments[0].forEach(e => {
@@ -34,7 +34,7 @@ String.prototype.format = function() {
   }
   return [...arguments].reduce((p, c) => p.replace(/%s/, c), this)
 }
-String.prototype.formatKey = function() {
+String.prototype.formatKey = function () {
   var formatted = this
   for (var prop in arguments[0]) {
     var regexp = new RegExp('\\{' + prop + '\\}', 'gi')
@@ -43,10 +43,37 @@ String.prototype.formatKey = function() {
   return formatted
 }
 if (!String.prototype.formatNumber) {
-  String.prototype.formatNumber = function() {
+  String.prototype.formatNumber = function () {
     var args = arguments
-    return this.replace(/{(\d+)}/g, function(match, number) {
+    return this.replace(/{(\d+)}/g, function (match, number) {
       return typeof args[number] !== 'undefined' ? args[number] : match
     })
   }
+}
+
+/**
+ * Format bytes as human-readable text.
+ *
+ * @param bytes Number of bytes.
+ * @param si True to use metric (SI) units, aka powers of 1000. False to use
+ *           binary (IEC), aka powers of 1024.
+ * @param dp Number of decimal places to display.
+ *
+ * @return Formatted string.
+ */
+function humanFileSize (bytes, si = false, dp = 1) {
+  const thresh = si ? 1000 : 1024
+
+  if (Math.abs(bytes) < thresh) return bytes + ' B'
+
+  const units = si ? ['kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'] : ['KiB', 'MiB', 'GiB', 'TiB', 'PiB', 'EiB', 'ZiB', 'YiB']
+  let u = -1
+  const r = 10 ** dp
+
+  do {
+    bytes /= thresh
+    ++u
+  } while (Math.round(Math.abs(bytes) * r) / r >= thresh && u < units.length - 1)
+
+  return bytes.toFixed(dp) + ' ' + units[u]
 }
